+++ libs/react-components/src/community/components/map-components/edit-actions/editing-actions.tsx
@@ -1,0 +1,14 @@
+import { JSX } from 'react';
+
+import styles from './editing-actions.module.scss';
+
+interface EditingActionsProps {
+  children?: React.ReactNode | React.ReactNode[];
+}
+
+const EditingActions = (props: EditingActionsProps): JSX.Element => {
+  const { children } = props;
+  return <div className={styles['tedi-editing-actions']}>{children}</div>;
+};
+
+export default EditingActions;